Moved credential types to mixin Added validation function for rotation type prop
